feat(db): expose physical scalar capabilities and preserve char semantics

+++ packages/libs/db/tests/integration/schema/inspector.test.ts
@@ -429,7 +429,7 @@ describeIntegrationDatabases('schema inspector', (context) => {
       expect(
         result?.columns.find((column) => column.columnName === 'flag'),
       ).toMatchObject({
-        dataType: 'integer',
+        dataType: 'decimal',
         nativeType: 'NUMBER(1,0)',
       });
       expect(
